requirements.txt
streamlit
pandas
numpy
plotly
pyngrok
python-dotenv
